[ADD/#93] 이미지 다운로드 구현
